requirements.txt
fastapi>=0.115.8
uvicorn>=0.34.0
pydantic>=2.10.6
openai>=1.61.1
requests>=2.32.3
python-multipart>=0.0.20
pyyaml>=6.0.1
gymnasium>=1.0.0
stable-baselines3>=2.8.0
shimmy>=2.0.0
matplotlib>=3.10.1
numpy>=2.2.3
openenv-core>=0.2.3
huggingface-hub>=0.29.1
